+++ frameworks/remix-run/application/playwright.config.ts
@@ -44,18 +44,18 @@ const config: PlaywrightTestConfig = {
                 ...devices['Desktop Chrome'],
             },
         },
-        // {
-        //     name: 'firefox',
-        //     use: {
-        //         ...devices['Desktop Firefox'],
-        //     },
-        // },
-        // {
-        //     name: 'webkit',
-        //     use: {
-        //         ...devices['Desktop Safari'],
-        //     },
-        // },
+        {
+            name: 'firefox',
+            use: {
+                ...devices['Desktop Firefox'],
+            },
+        },
+        {
+            name: 'webkit',
+            use: {
+                ...devices['Desktop Safari'],
+            },
+        },
 
         /* Test against mobile viewports. */
         // {
